requirements.txt
beautifulsoup4
requests
googlesearch-python
selenium
webdriver-manager
gspread
google-auth